회원가입1차수정
