+++ src/main/resources/templates/company_main.html
@@ -65,30 +65,30 @@
         <ul class="pagination">
           <!-- 처음 버튼 -->
           <li th:classappend="${paging.firstPageBtn ? '' : 'disabled'}">
-            <a class="page-link" th:href="@{|/?page=1|}" aria-label="처음">&laquo;</a>
+            <a class="page-link" th:href="@{|/company?page=1|}" aria-label="처음">&laquo;</a>
           </li>
 
           <!-- 이전 버튼 -->
           <li th:classappend="${paging.prevBtn ? '' : 'disabled'}">
-            <a class="page-link" th:href="@{|/?page=${paging.startPage - 1}|}"
+            <a class="page-link" th:href="@{|/company?page=${paging.startPage - 1}|}"
                aria-label="이전">&lt;</a>
           </li>
 
           <!-- 페이지 번호 -->
           <li th:each="i : ${#numbers.sequence(paging.startPage, paging.endPage)}"
               th:classappend="${paging.currentPage == i ? 'active' : ''}">
-            <a class="page-link" th:href="@{|/?page=${i}|}" th:text="${i}"></a>
+            <a class="page-link" th:href="@{|/company?page=${i}|}" th:text="${i}"></a>
           </li>
 
           <!-- 다음 버튼 -->
           <li th:classappend="${paging.nextBtn ? '' : 'disabled'}">
-            <a class="page-link" th:href="@{|/?page=${paging.endPage + 1}|}"
+            <a class="page-link" th:href="@{|/company?page=${paging.endPage + 1}|}"
                aria-label="다음">&gt;</a>
           </li>
 
           <!-- 마지막 버튼 -->
           <li th:classappend="${paging.lastPageBtn ? '' : 'disabled'}">
-            <a class="page-link" th:href="@{|/?page=${paging.totalPage}|}"
+            <a class="page-link" th:href="@{|/company?page=${paging.totalPage}|}"
                aria-label="마지막">&raquo;</a>
           </li>
         </ul>
